requirements.txt
customtkinter>=5.2.1
opencv-python>=4.8.1
pygetwindow>=0.0.9
pynput>=1.7.6
Pillow>=10.0.1
numpy>=1.26.0
winsound; sys_platform == 'win32'